+++ CHANGES.md
@@ -4,6 +4,10 @@ moodle-tool_coursedates
 Changes
 -------
 
+### Unreleased
+
+* 2020-01-13 - Improvement: Only add the category navigation node if there are existing course fields.
+
 ### v3.7-r1
 
 * 2019-10-17 - Initial version.